requirements.txt
django==4.0.4
Pillow==9.1.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
python-dotenv==0.20.0